chore: Rename the v3federation to v4federation
Keystone-ng moved the new auth method to the v4 (with passkeys and the
jwt login). Start working on support for those new features by first
moving the corresponding federation login.
