+++ openstack_sdk/src/auth.rs
@@ -31,20 +31,20 @@ mod auth_token_endpoint;
 pub mod authtoken;
 pub mod authtoken_scope;
 pub mod v3applicationcredential;
-#[cfg(feature = "keystone_ng")]
-pub mod v3federation;
 pub mod v3oidcaccesstoken;
 pub mod v3password;
 pub mod v3token;
 pub mod v3totp;
 pub mod v3websso;
+#[cfg(feature = "keystone_ng")]
+pub mod v4federation;
 
 use authtoken::{AuthToken, AuthTokenError};
 use authtoken_scope::AuthTokenScopeError;
-#[cfg(feature = "keystone_ng")]
-use v3federation::FederationError;
 use v3oidcaccesstoken::OidcAccessTokenError;
 use v3websso::WebSsoError;
+#[cfg(feature = "keystone_ng")]
+use v4federation::FederationError;
 
 /// Authentication error
 #[derive(Debug, Error)]
@@ -96,7 +96,7 @@ impl From<WebSsoError> for AuthError {
 
 #[cfg(feature = "keystone_ng")]
 impl From<FederationError> for AuthError {
-    fn from(source: v3federation::FederationError) -> Self {
+    fn from(source: v4federation::FederationError) -> Self {
         Self::AuthToken {
             source: source.into(),
         }
